requirements.txt
torch
numpy
pandas
PyYAML
matplotlib
tqdm
nuscenes-devkit
scikit-learn
tensorboard